+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.0.1
+* Init
 
-* TODO: Describe initial release.
+## 0.0.2
+* Added Logger
+* Added config option
+* Few API improvement
+* Added MonitoredWidget which let you monitor time from init to first widget built. (example case: if you want to measure time between loading and rendered page)
+* enterAction now can be called with parent
+* MonitoredHttpClient can now be entered with parent action